[upgrade] #43 | task1 Split codes. Impl a class.
