+++ LuminanceLimiterSG/src/luminance_limiter_sg.h
@@ -25,7 +25,7 @@ namespace luminance_limiter_sg {
 		NormalizedY minimum() const noexcept;
 
 		template<typename F>
-		BOOL pixelwise_map(const F&& f) {
+		BOOL pixelwise_map(const F& f) {
 			for (auto& elem : this->buffer)
 			{
 				elem = f(elem);
